Add files via upload
This is an application for my personal Minecraft server. The purpose of this program is to provide players with easier access to information about the server, login accounts, and quick access to the website. Of course, I also wish to create an application for myself. This is my first time using C++ to develop a Windows application.If you wish to run this program (Note 14: it is currently in the early development stage), you can simply download and open the executable file released by Life [Important: Do not use the DEBUG version].